project-api/requirements.txt
alembic==1.16.2
email_validator==2.2.0
fastapi==0.115.13
fastapi-pagination==0.13.2
Jinja2==3.1.6
passlib[bcrypt]==1.7.4
postmarker==1.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.4
sentry-sdk==2.30.0
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
tenacity==9.1.2
uvicorn[standard]==0.34.3
